+++ src/components/DappBrowser/search/bar/AccountIcon.tsx
@@ -7,7 +7,7 @@ import { getAccountProfileInfo } from '@/helpers/accountInfo';
 import Routes from '@/navigation/routesNames';
 import { ContactAvatar } from '@/components/contacts';
 import { Bleed } from '@/design-system';
-import { ContextMenuButton } from '../../context-menu';
+import { ContextMenuButton } from '../../../context-menu';
 import { Network } from '@/networks/types';
 
 import { RainbowNetworks, getNetworkObj } from '@/networks';
@@ -16,8 +16,8 @@ import store from '@/redux/store';
 import { showActionSheetWithOptions } from '@/utils';
 import * as i18n from '@/languages';
 import { useAppSessionsStore } from '@/state/appSessions';
-import { useBrowserContext } from '../BrowserContext';
-import { getDappHost } from '../handleProviderRequest';
+import { useBrowserContext } from '../../BrowserContext';
+import { getDappHost } from '../../handleProviderRequest';
 import { Address, toHex } from 'viem';
 import { handleDappBrowserConnectionPrompt } from '@/utils/requestNavigationHandlers';
 import { getNetworkFromChainId } from '@/utils/ethereumUtils';
